refactor the controller and subscriber
