+++ inc/Engine/Optimization/RUCSS/Controller/UsedCSS.php
@@ -3,10 +3,13 @@ declare(strict_types=1);
 
 namespace WP_Rocket\Engine\Optimization\RUCSS\Controller;
 
+use WP_Rocket\Admin\Options_Data;
 use WP_Rocket\Engine\Cache\Purge;
 use WP_Rocket\Engine\Optimization\RegexTrait;
 use WP_Rocket\Engine\Optimization\RUCSS\Database\Row\UsedCSS as UsedCSS_Row;
 use WP_Rocket\Engine\Optimization\RUCSS\Database\Queries\UsedCSS as UsedCSS_Query;
+use WP_Rocket\Engine\Optimization\RUCSS\Frontend\APIClient;
+use WP_Rocket\Logger\Logger;
 
 class UsedCSS {
 	use RegexTrait;
@@ -26,21 +29,133 @@ class UsedCSS {
 	private $purge;
 
 	/**
-	 * Status of CPCSS option.
+	 * Plugin options instance.
 	 *
-	 * @var bool
+	 * @var Options_Data
 	 */
-	private $cpcss_enabled = false;
+	protected $options;
 
 	/**
-	 * Instantiate the class
+	 * APIClient instance
 	 *
-	 * @param UsedCSS_Query $used_css_query UsedCSS Query instance.
-	 * @param Purge         $purge          Purge instance.
+	 * @var APIClient
 	 */
-	public function __construct( UsedCSS_Query $used_css_query, Purge $purge ) {
+	private $api;
+
+
+	/**
+	 * Instantiate the class.
+	 *
+	 * @param Options_Data  $options Options instance.
+	 * @param UsedCSS_Query $used_css_query Usedcss Query instance.
+	 * @param Purge         $purge Purge instance.
+	 * @param APIClient     $api Apiclient instance.
+	 */
+	public function __construct( Options_Data $options, UsedCSS_Query $used_css_query, Purge $purge, APIClient $api ) {
+		$this->options        = $options;
 		$this->used_css_query = $used_css_query;
 		$this->purge          = $purge;
+		$this->api            = $api;
+	}
+
+	/**
+	 * Determines if we treeshake the CSS.
+	 *
+	 * @return boolean
+	 */
+	public function is_allowed() : bool {
+		if ( rocket_get_constant( 'DONOTROCKETOPTIMIZE' ) ) {
+			return false;
+		}
+
+		if ( rocket_bypass() ) {
+			return false;
+		}
+
+		if ( ! (bool) $this->options->get( 'remove_unused_css', 0 ) ) {
+			return false;
+		}
+
+		// Bailout if user is logged in and cache for logged in customers is active.
+		if ( is_user_logged_in() && (bool) $this->options->get( 'cache_logged_user', 0 ) ) {
+			return false;
+		}
+
+		return true;
+	}
+
+	/**
+	 * Apply TreeShaked CSS to the current HTML page.
+	 *
+	 * @param string $html  HTML content.
+	 *
+	 * @return string  HTML content.
+	 */
+	public function treeshake( string $html ) : string {
+		if ( ! $this->is_allowed() ) {
+			return $html;
+		}
+
+		global $wp;
+		$url       = untrailingslashit( home_url( add_query_arg( [], $wp->request ) ) );
+		$is_mobile = $this->is_mobile();
+		$used_css  = $this->get_used_css( $url, $is_mobile );
+
+		if ( empty( $used_css ) || ( $used_css->retries < 3 ) ) {
+			$config = [
+				'treeshake'      => 1,
+				'wpr_email'      => $this->options->get( 'consumer_email', '' ),
+				'wpr_key'        => $this->options->get( 'consumer_key', '' ),
+				'rucss_safelist' => $this->options->get( 'remove_unused_css_safelist', [] ),
+			];
+
+			$treeshaked_result = $this->api->optimize( $html, $url, $config );
+
+			if ( 200 !== $treeshaked_result['code'] ) {
+				Logger::error(
+					'Error when contacting the RUCSS API.',
+					[
+						'rucss error',
+						'url'     => $url,
+						'code'    => $treeshaked_result['code'],
+						'message' => $treeshaked_result['message'],
+					]
+				);
+				return $html;
+			}
+
+			$retries = 0;
+			if ( isset( $used_css->retries ) ) {
+				$retries = $used_css->retries;
+			}
+
+			if ( ! empty( $treeshaked_result['unprocessed_css'] ) ) {
+				$this->schedule_rucss_retry();
+			}
+
+			$data = [
+				'url'            => $url,
+				'css'            => $treeshaked_result['css'],
+				'unprocessedcss' => wp_json_encode( $treeshaked_result['unprocessed_css'] ),
+				'retries'        => empty( $treeshaked_result['unprocessed_css'] ) ? 3 : $retries + 1,
+				'is_mobile'      => $is_mobile,
+				'modified'       => current_time( 'mysql', true ),
+			];
+
+			$used_css = $this->save_or_update_used_css( $data );
+
+			if ( ! $used_css ) {
+				return $html;
+			}
+		}
+
+		$html = $this->remove_used_css_from_html( $html, $used_css->unprocessedcss );
+
+		$html = $this->add_used_css_to_html( $html, $used_css );
+
+		$this->update_last_accessed( (int) $used_css->id );
+
+		return $html;
 	}
 
 	/**
@@ -72,6 +187,10 @@ class UsedCSS {
 	 * @return void
 	 */
 	public function retries_pages_with_unprocessed_css() {
+		if ( ! (bool) $this->options->get( 'remove_unused_css', 0 ) ) {
+			return;
+		}
+
 		$used_css_list = $this->get_used_css_with_unprocessed_css();
 
 		foreach ( $used_css_list as $used_css_item ) {
@@ -238,17 +357,15 @@ class UsedCSS {
 	/**
 	 * Alter HTML string and add the used CSS style in <head> tag,
 	 *
-	 * @param string $html HTML content.
-	 * @param string $used_css Used CSS content.
-	 * @param string $url Page Url.
-	 * @param bool   $is_mobile For mobile or desktop.
+	 * @param string      $html HTML content.
+	 * @param UsedCSS_Row $used_css Used CSS DB row.
 	 *
 	 * @return string HTML content.
 	 */
-	public function add_used_css_to_html( string $html, string $used_css, $url = '', bool $is_mobile = false ) : string {
+	public function add_used_css_to_html( string $html, UsedCSS_Row $used_css ) : string {
 		return preg_replace(
 			'#</title>#iU',
-			'</title>' . $this->get_used_css_markup( $used_css, $url, $is_mobile ),
+			'</title>' . $this->get_used_css_markup( $used_css ),
 			$html,
 			1
 		);
@@ -315,26 +432,15 @@ class UsedCSS {
 	}
 
 	/**
-	 * Set status of CPCSS option
-	 *
-	 * @param bool $status Status of CPCSS option.
-	 */
-	public function set_cpcss_enabled( bool $status ) {
-		$this->cpcss_enabled = (bool) $status;
-	}
-
-	/**
 	 * Save Used CSS into filesystem in case CPCSS is enabled only.
 	 *
-	 * @param string $url Page Url.
-	 * @param string $used_css Used CSS contents.
-	 * @param bool   $is_mobile Used CSS for mobile or desktop.
+	 * @param UsedCSS_Row $used_css Used CSS contents.
 	 *
 	 * @return bool
 	 */
-	private function save_used_css_in_filesystem( string $url, string $used_css, bool $is_mobile = false ) : bool {
+	private function save_used_css_in_filesystem( UsedCSS_Row $used_css ) : bool {
 
-		if ( ! $this->cpcss_enabled ) {
+		if ( ! $this->cpcss_enabled() ) {
 			return false;
 		}
 
@@ -346,9 +452,9 @@ class UsedCSS {
 			}
 		}
 
-		$used_css_filepath = rocket_get_constant( 'WP_ROCKET_USED_CSS_PATH' ) . $this->get_used_css_filename( $url, $is_mobile );
+		$used_css_filepath = rocket_get_constant( 'WP_ROCKET_USED_CSS_PATH' ) . $this->get_used_css_filename( $used_css->url, (bool) $used_css->is_mobile );
 
-		return rocket_put_content( $used_css_filepath, $used_css );
+		return rocket_put_content( $used_css_filepath, $used_css->css );
 	}
 
 	/**
@@ -366,30 +472,65 @@ class UsedCSS {
 	/**
 	 * Return Markup for used_css into the page.
 	 *
-	 * @param string $used_css Used CSS contents.
-	 * @param string $url Page url.
-	 * @param bool   $is_mobile Used CSS for mobile or desktop.
+	 * @param UsedCSS_Row $used_css Used CSS DB Row.
 	 *
 	 * @return string
 	 */
-	private function get_used_css_markup( string $used_css, string $url = '', bool $is_mobile = false ) : string {
-		if ( ! $this->cpcss_enabled ) {
+	private function get_used_css_markup( UsedCSS_Row $used_css ) : string {
+
+		if ( ! $this->cpcss_enabled() ) {
 			return sprintf(
 				'<style id="wpr-usedcss">%s</style>',
-				wp_strip_all_tags( $used_css )
+				wp_strip_all_tags( $used_css->css )
 			);
 		}
 
-		$used_css_filename = $this->get_used_css_filename( $url, $is_mobile );
+		$used_css_filename = $this->get_used_css_filename( $used_css->url, (bool) $used_css->is_mobile );
 		$used_css_filepath = rocket_get_constant( 'WP_ROCKET_USED_CSS_PATH' ) . $used_css_filename;
 
 		if ( ! rocket_direct_filesystem()->exists( $used_css_filepath ) ) {
-			$this->save_used_css_in_filesystem( $url, $used_css, $is_mobile );
+			$this->save_used_css_in_filesystem( $used_css );
 		}
 
 		return sprintf(
 			'<link rel="stylesheet" id="wpr-usedcss-css" href="%s">', // phpcs:ignore WordPress.WP.EnqueuedResources.NonEnqueuedStylesheet
 			rocket_get_constant( 'WP_ROCKET_USED_CSS_URL' ) . $used_css_filename
 		);
+	}
+
+	/**
+	 * Determines if the page is mobile and separate cache for mobile files is enabled.
+	 *
+	 * @return boolean
+	 */
+	private function is_mobile() : bool {
+		return $this->options->get( 'cache_mobile', 0 ) &&
+			$this->options->get( 'do_caching_mobile_files', 0 ) &&
+			wp_is_mobile();
+	}
+
+	/**
+	 * Schedules RUCSS to retry pages with missing CSS files.
+	 * Retries happen after 30 minutes.
+	 *
+	 * @return void
+	 */
+	public function schedule_rucss_retry() {
+		$scheduled = wp_next_scheduled( 'rocket_rucss_retries_cron' );
+
+		if ( $scheduled ) {
+			return;
+		}
+
+		wp_schedule_single_event( time() + ( 0.5 * HOUR_IN_SECONDS ), 'rocket_rucss_retries_cron' );
+	}
+
+	/**
+	 * Is CPCSS option active or not?
+	 *
+	 * @return bool
+	 */
+	private function cpcss_enabled() {
+		return (bool) $this->options->get( 'async_css', 0 );
 	}
 }
